requirements.txt
django==2.1.5
djangorestframework==3.8.2
libsass==0.15.1
python-decouple==3.1
flake8==3.7.6
isort==4.3.4
